#38: Implement custom auth headers. Reworked some code.
